refactored auth routes
